new query is added
